requirements.txt
# Flask Core
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Security
bcrypt==4.1.2
python-dotenv==1.0.0

# Utilities
email-validator==2.1.0
Werkzeug==3.0.1
requests==2.31.0
httpx==0.27.0

# OpenAI Integration
openai==1.12.0

# Background Task Scheduling
APScheduler==3.10.4

# Image Processing
Pillow==10.2.0
split-image==2.0.1

# S3 Storage
boto3==1.34.34
botocore==1.34.34
